Removed unused variables and imports and moved common attributes to Autonomous class
