+++ Autonomous.java
@@ -38,16 +38,6 @@ public abstract class Autonomous {
 		}
 	}
 	
-	String GameLayout = DriverStation.getInstance().getGameSpecificMessage();
-	
-	protected Timer timer;
-	protected Claw claw;
-	protected Tilt tilt;
-	protected Elevator elevator;
-	
-	protected double horizontal, vertical, rotation;
-	Gyro gyro;
-	
 	protected enum Action {
 		TILT_FORWARD,
 		CALIBRATE_ELEVATOR,
@@ -58,9 +48,39 @@ public abstract class Autonomous {
 		RAISE_ELEVATOR,
 		DROP_POWERCUBE,
 		STOP;
+	}	
+	protected static class Movement {
+		public final double SPEED;
+		public final double ANGLE_START;
+		public final double ANGLE_END;
+		public final double TIMEOUT;
+		
+		public Movement(double speedIn, double angleStartIn, double angleEndIn, double timeoutIn) {
+			SPEED = speedIn;
+			ANGLE_START = angleStartIn;
+			ANGLE_END = angleEndIn;
+			TIMEOUT = timeoutIn;
+		}
 	}
 	
-	Action action;
+	String GameLayout = DriverStation.getInstance().getGameSpecificMessage();
+	
+	protected Timer timer;
+	protected Claw claw;
+	protected Tilt tilt;
+	protected Elevator elevator;
+	
+	protected double horizontal, vertical, rotation;
+	Gyro gyro;
+	
+	protected double speed;
+	protected double angleStart;
+	protected double angleEnd;
+	protected double angle;
+	protected double timeout;
+	
+	
+	protected Action action;
 	protected SwitchState switchState;
 	
 	public Autonomous(Gyro gyroIn, SwitchState switchStateIn) {
